requirements.txt
asgiref==3.8.1
certifi==2025.7.14
Django==5.2.1
django-extensions==4.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
